pyproject.toml
[project]
name = "maou"
version = "0.2.0"
description = "shogi ai"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
license = {text = "GPL-3.0-only"}
requires-python = ">=3.9,<3.12"
dependencies = [
    "click>=8.1.7",
    "cshogi>=0.9.2",
    "tqdm>=4.67.1",
    "pyarrow>=19.0.0",
    "numpy>=1.19.0",
    "google-crc32c>=1.6.0",
    "psutil>=6.0.0",
    "pynvml>=11.4.1",
]

[project.optional-dependencies]
# GPU種類ごとのextra
cpu = [
    "torch>=2.6.0,<=2.7.0",
    "torchinfo>=1.8.0",
    "torch-tb-profiler>=0.4.3",
]
cuda = [
    "torch>=2.6.0,<=2.7.0",
    "torchinfo>=1.8.0",
    "torch-tb-profiler>=0.4.3",
]
mpu = [
    "torch>=2.6.0,<=2.7.0",
    "torchinfo>=1.8.0",
    "torch-tb-profiler>=0.4.3",
]
tpu = [
    "torch-xla>=2.6.0,<=2.7.0",
    "torch>=2.6.0,<=2.7.0",
    "torchinfo>=1.8.0",
    "torch-tb-profiler>=0.4.3",
]
# クラウドプロバイダーごとのextra
gcp = [
    "google-cloud-storage>=2.19.0",
    "google-cloud-bigquery[pandas]>=3.27.0",
    "google-cloud-bigquery-storage>=2.28.0",
]
aws = [
    "boto3>=1.34.0",
]

[project.scripts]
maou = "maou.infra.console.app:main"

[tool.poetry]
packages = [{include = "maou", from = "src"}]

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.2.0"
flake8 = "^7.2.0"
yamllint = "^1.37.1"
isort = "^6.0.1"
commitizen = "^4.8.2"
pytest = "^8.3.5"
ruff = "^0.11.12"
types-tqdm = "^4.67.0.20241221"
mypy = "^1.15.0"
google-cloud-storage = "^2.19.0"
google-cloud-bigquery = {version = "^3.27.0", extras = ["pandas"]}
google-cloud-bigquery-storage = "^2.28.0"
boto3 = "^1.34.0"
autopep8 = "^2.3.2"
yapf = "^0.43.0"

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "pep440"
version_provider = "poetry"
update_changelog_on_bump = true
major_version_zero = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]

[tool.mypy]
disallow_untyped_defs = true
plugins = [
  "numpy.typing.mypy_plugin",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
  "click",
  "cshogi",
  "pytest",
  "torchinfo",
  "pyarrow.*",
  "google_crc32c",
  "google.cloud",
  "maou.*",
]

[tool.isort]
profile = "black"

[tool.ruff]
line-length = 64
indent-width = 4

[[tool.poetry.source]]
name = "pytorch-cuda"
url = "https://download.pytorch.org/whl/cu118"
priority = "explicit"
